Automatically Load New .WEA and .CRT
I hex-edited new CRTs and WEAs for anything that didn't have them, and now it will load those files. the weapons ones are being worked on but can be forced into your inventory if you edit Default_PC.CRT, but so far I can't get any of them to be equipped.